+++ tests/browser/evaluate.test.js
@@ -10,7 +10,7 @@ describe("Evaluate", function() {
 
     let browser;
     before(async() => {
-        browser = await Wendigo.createBrowser({log: true});
+        browser = await Wendigo.createBrowser();
     });
 
     beforeEach(async() => {